Change delete product to take space rather than uuid.
