+++ ui/src/Products/ProductClient.test.tsx
@@ -1,28 +1,16 @@
-/*
- * Copyright (c) 2020 Ford Motor Company
- * All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
+import ProductClient from './ProductClient';
+import {Product} from './Product';
+import axios, {AxiosResponse} from 'axios';
+import {MatomoWindow} from '../CommonTypes/MatomoWindow';
+import TestUtils from "../tests/TestUtils";
+import Cookies from "universal-cookie";
+import Axios from "axios";
 
-import Axios, {AxiosResponse} from 'axios';
-import ProductClient from '../Products/ProductClient';
-import TestUtils from '../tests/TestUtils';
-import Cookies from 'universal-cookie';
+jest.mock('axios');
+declare let window: MatomoWindow;
 
 describe('Product Client', function() {
-    const spaceUuid = 'uuid';
-    const baseProductsUrl = `/api/spaces/${spaceUuid}/products`;
+    const baseProductsUrl = `/api/spaces/${TestUtils.space.uuid}/products`;
     const expectedConfig = {
         headers: {
             'Content-Type': 'application/json',
@@ -64,7 +52,7 @@ describe('Product Client', function() {
 
     it('should update a product and return that product', function(done) {
         const expectedUrl = `${baseProductsUrl}/${TestUtils.productWithAssignments.id}`;
-        ProductClient.editProduct(spaceUuid, TestUtils.productWithAssignments)
+        ProductClient.editProduct(TestUtils.space.uuid!!, TestUtils.productWithAssignments)
             .then((response) => {
                 expect(Axios.put).toHaveBeenCalledWith(expectedUrl, TestUtils.productWithAssignments, expectedConfig);
                 expect(response.data).toBe('Updated Product');
@@ -74,7 +62,7 @@ describe('Product Client', function() {
 
     it('should delete a product', function(done) {
         const expectedUrl = `${baseProductsUrl}/${TestUtils.productWithAssignments.id}`;
-        ProductClient.deleteProduct(spaceUuid, TestUtils.productWithAssignments)
+        ProductClient.deleteProduct(TestUtils.space, TestUtils.productWithAssignments)
             .then((response) => {
                 expect(Axios.delete).toHaveBeenCalledWith(expectedUrl, expectedConfig);
                 expect(response.data).toBe('Deleted Product');
@@ -85,7 +73,7 @@ describe('Product Client', function() {
     it('should return the products given a date', function(done) {
         const date = '2019-01-10';
         const expectedUrl = baseProductsUrl + `?requestedDate=${date}`;
-        ProductClient.getProductsForDate(spaceUuid, new Date(2019, 0, 10))
+        ProductClient.getProductsForDate(TestUtils.space.uuid!!, new Date(2019, 0, 10))
             .then((response) => {
                 expect(Axios.get).toHaveBeenCalledWith(expectedUrl, expectedConfig);
                 expect(response.data).toBe('Get Products');
@@ -93,4 +81,64 @@ describe('Product Client', function() {
             });
 
     });
+
+    describe('Matomo', () => {
+
+        let originalWindow: Window;
+
+        beforeEach(() => {
+            originalWindow = window;
+        });
+
+        afterEach(() => {
+            (window as Window) = originalWindow;
+        });
+
+        it('should push create product action on create', async () => {
+            const expectedName = 'Floam';
+            const product: Product = {
+                archived: false,
+                assignments: [],
+                id: 0,
+                name: expectedName,
+                productTags: [],
+                spaceId: 0,
+            };
+
+            const expectedResponse = {};
+            axios.post = jest.fn(() => Promise.resolve(expectedResponse as any));
+
+            const axiosResponse = await ProductClient.createProduct(TestUtils.space, product);
+
+            expect(axiosResponse).toBe(expectedResponse);
+
+            expect(window._paq).toContainEqual(['trackEvent', TestUtils.space.name, 'createProduct', expectedName]);
+
+        });
+
+        it('should push createError on create with failure code', async () => {
+            const expectedName = 'Floam';
+            const product: Product = {
+                archived: false,
+                assignments: [],
+                id: 0,
+                name: expectedName,
+                productTags: [],
+                spaceId: 0,
+            };
+
+            const expectedResponse = {code: 417};
+            axios.post = jest.fn(() => Promise.reject(expectedResponse as any));
+
+            try {
+                await ProductClient.createProduct(TestUtils.space, product);
+                fail('Did not catch an exception from the product creation');
+            } catch (err)  {
+                expect(window._paq).toContainEqual(['trackEvent', TestUtils.space.name, 'createProductError', expectedName, expectedResponse.code]);
+            }
+
+        });
+    });
+
 });
+
