+++ ui/src/Products/ProductClient.ts
@@ -60,8 +60,8 @@ class ProductClient {
         return Axios.put(url, product, config);
     }
 
-    static async deleteProduct(spaceUuid: string, product: Product): Promise<AxiosResponse> {
-        const url = this.getBaseProductsUrl(spaceUuid) + `/${product.id}`;
+    static async deleteProduct(space: Space, product: Product): Promise<AxiosResponse> {
+        const url = this.getBaseProductsUrl(space.uuid!!) + `/${product.id}`;
         const config = {
             headers: {
                 'Content-Type': 'application/json',
