+++ ui/src/tests/Product.test.tsx
@@ -470,7 +470,7 @@ describe('Products', () => {
                 fireEvent.click(deleteButton);
             });
             expect(ProductClient.deleteProduct).toBeCalledTimes(1);
-            expect(ProductClient.deleteProduct).toBeCalledWith(TestUtils.space.uuid, TestUtils.productWithoutAssignments);
+            expect(ProductClient.deleteProduct).toBeCalledWith(TestUtils.space, TestUtils.productWithoutAssignments);
         });
 
         it('should not show archive button option in delete modal if product is already archived', async () => {
